requirements.txt
pytube
clilib
requests
flask
flask-session
gunicorn
ffmpeg-python
redis
plexapi